+++ src/mimic_base/src/types/bytes.rs
@@ -8,13 +8,13 @@ use mimic::orm::prelude::*;
 #[newtype(
     primitive = "Blob",
     value(item(is = "types::Blob")),
-    traits(remove(Validate))
+    traits(remove(ValidateManual))
 )]
 pub struct Bytes<const LEN: usize> {}
 
 #[allow(clippy::cast_possible_wrap)]
-impl<const LEN: usize> Validate for Bytes<LEN> {
-    fn validate(&self) -> Result<(), ErrorVec> {
+impl<const LEN: usize> ValidateManual for Bytes<LEN> {
+    fn validate_manual(&self) -> Result<(), ErrorVec> {
         let errs = ErrorVec::from_result(validator::len::Max::validate(
             &self.0.to_vec(),
             LEN as isize,
@@ -31,13 +31,13 @@ impl<const LEN: usize> Validate for Bytes<LEN> {
 #[newtype(
     primitive = "Blob",
     value(item(is = "types::Blob")),
-    traits(remove(Validate))
+    traits(remove(ValidateManual))
 )]
 pub struct Utf8<const LEN: usize> {}
 
 #[allow(clippy::cast_possible_wrap)]
-impl<const LEN: usize> Validate for Utf8<LEN> {
-    fn validate(&self) -> Result<(), ErrorVec> {
+impl<const LEN: usize> ValidateManual for Utf8<LEN> {
+    fn validate_manual(&self) -> Result<(), ErrorVec> {
         let mut errs = ErrorVec::default();
 
         // utf8