+++ src/mimic_base/src/types/text.rs
@@ -10,13 +10,13 @@ use mimic::orm::prelude::*;
 #[newtype(
     primitive = "String",
     value(item(is = "types::String")),
-    traits(add(Hash), remove(Validate))
+    traits(add(Hash), remove(ValidateManual))
 )]
 pub struct Text<const LEN: usize> {}
 
 #[allow(clippy::cast_possible_wrap)]
-impl<const LEN: usize> Validate for Text<LEN> {
-    fn validate(&self) -> Result<(), ErrorVec> {
+impl<const LEN: usize> ValidateManual for Text<LEN> {
+    fn validate_manual(&self) -> Result<(), ErrorVec> {
         ErrorVec::from_result(validator::len::Max::validate(&self.0, LEN as isize)).result()
     }
 }